make the request validator do what the response validator does with data validation
