+++ src/Validation/RequestValidator.php
@@ -10,17 +10,22 @@ use Spectator\Exceptions\RequestValidationException;
 
 class RequestValidator
 {
+    use SchemaValidator;
+
     protected $request;
 
     protected $pathItem;
 
     protected $method;
 
-    public function __construct(Request $request, PathItem $pathItem, $method)
+    protected $version;
+
+    public function __construct(Request $request, PathItem $pathItem, $method, $version = '3.0')
     {
         $this->request = $request;
         $this->pathItem = $pathItem;
         $this->method = strtolower($method);
+        $this->version = $version;
     }
 
     public static function validate(Request $request, PathItem $pathItem, $method)
@@ -120,7 +125,7 @@ class RequestValidator
             }
         }
 
-        $result = $validator->validate($body, $jsonSchema->getSerializableData());
+        $result = $validator->validate($body, $this->prepareData($jsonSchema->getSerializableData()));
 
         if (! $result->isValid()) {
             throw RequestValidationException::withError('Request body did not match provided JSON schema.', $result->error());
